Update to Airflow 3 API route
Once https://github.com/blue-core-lod/bluecore-workflows/pull/41 is
merged to upgrade bluecore-workflows to Airflow 3 bluecore_api will need
to talk to the new Airflow API route to trigger the DAG. It appears that
the route is the same except for the api version number:

https://airflow.apache.org/docs/apache-airflow/stable/stable-rest-api-ref.html#operation/trigger_dag_run

There are a lot of changes here related to moving the `src/bluecore`
directory to `src/bluecore_api`. Also, it adds a `bluecore` command line
utility which you can use to get an access token, and load jsonld data.

The documentation was updated significantly to reflect that to develop
this code you will want to checkout bluecore-workflows, and run that
`docker compose up` in order to get the Postgre, Airflow and Keycloak
services up that bluecore_api depends on.