Upgrade bluecore-models to v0.8.1
This commit includes changes to start using v0.8.1 which introduced
automatic framing of JSON-LD data when assigning to Work.data or
Instance.data. This meant that use of
`bluecore_models.utils.graph.frame_jsonld()` could be removed from
bluecore_api.

The change did require some changes to test fixture data in how Work and
Instance objects are created and modified.

The `.data` property must be set after `.uri` when instantiating Work and
Instance objects. This is because framing requires that `.uri` already
be set.

Also, the JSON-LD data used needs to be legitimate JSON-LD when testing
the change_documents. For example simply using `{"name": "foo"}` will
get framed as an empty graph since it lacks a `@id` and `@type`. This in
itself doesn't cause an error, but the tests were expecting updates to
happen, but since the `.data` property wasn't changing no update was
needed.

Closes #125